Feature/book crud

+++ src/modules/category/category.service.ts
@@ -1,13 +1,16 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, Inject, forwardRef } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CategoryCreateUpdateDto } from './dto/category.create-update.dto';
 import { CategoryDto } from './dto/category.dto';
 import { plainToInstance } from 'class-transformer';
 import { CategoryValidator } from './category.validator';
+import { BookService } from '../book/book.service';
 
 @Injectable()
 export class CategoryService {
-  constructor(private readonly prisma: PrismaService, private readonly categoryValidator: CategoryValidator) {}
+  constructor(private readonly prisma: PrismaService, 
+              private readonly categoryValidator: CategoryValidator,
+              @Inject(forwardRef(() => BookService)) private readonly bookService: BookService) {}
 
   async createCategory({ name }: CategoryCreateUpdateDto): Promise<CategoryDto> {
     await this.categoryValidator.validateCategoryNameExists(name);
@@ -34,6 +37,9 @@ export class CategoryService {
 
   async deleteCategory(id: number): Promise<void> {
     await this.findCategoryById(id);
+    if((await this.bookService.findBooks({ categoryId: id })).length > 0) {
+      throw new HttpException(`Existem livros vinculados a categoria de id ${id}. Exclusão não é permitida`, HttpStatus.CONFLICT);
+    }
     await this.prisma.category.delete({ where: { id }});
     
     return;
